+++ myst_nb/execution.py
@@ -9,22 +9,21 @@ The primary methods in this module are:
   or if 'auto' / 'force' is set, will execute the notebook.
 
 """
-from datetime import datetime
 import os
+import re
 import tempfile
-from typing import List, Optional, Set
+from datetime import datetime
+from pathlib import Path
+from typing import Iterable, List, Optional, Set
 
 import nbformat as nbf
-from pathlib import Path
-
+from jupyter_cache import get_cache
+from jupyter_cache.executors import load_executor
+from jupyter_cache.executors.utils import single_nb_execution
 from sphinx.application import Sphinx
 from sphinx.builders import Builder
 from sphinx.environment import BuildEnvironment
 from sphinx.util import logging, progress_message
-
-from jupyter_cache import get_cache
-from jupyter_cache.executors import load_executor
-from jupyter_cache.executors.utils import single_nb_execution
 
 from .converter import get_nb_converter
 
@@ -227,8 +226,11 @@ def is_valid_exec_file(env: BuildEnvironment, docname: str) -> bool:
     doc_path = env.doc2path(docname)
     if doc_path in env.nb_excluded_exec_paths:
         return False
-    extension = os.path.splitext(doc_path)[1]
-    if extension not in env.nb_allowed_exec_suffixes:
+    matches = tuple(
+        re.search(re.escape(suffix) + "$", doc_path)
+        for suffix in env.nb_allowed_exec_suffixes
+    )
+    if not any(matches):
         return False
     return True
 
@@ -325,7 +327,9 @@ def execute_staged_nb(
     return result
 
 
-def nb_has_all_output(source_path: str, nb_extensions: List[str] = (".ipynb",)) -> bool:
+def nb_has_all_output(
+    source_path: str, nb_extensions: Iterable[str] = (".ipynb",)
+) -> bool:
     """Determine if the path contains a notebook with at least one output."""
     has_outputs = False
     ext = os.path.splitext(source_path)[1]