+++ tests/test_text_based.py
@@ -15,6 +15,7 @@ def test_basic_run(sphinx_run, file_regression, check_nbs):
         "author",
         "source_map",
         "language_info",
+        "wordcount",
     }
     assert sphinx_run.app.env.metadata["basic_unrun"]["author"] == "Chris"
     assert (
@@ -36,7 +37,6 @@ def test_basic_run(sphinx_run, file_regression, check_nbs):
 def test_basic_run_exec_off(sphinx_run, file_regression, check_nbs):
     sphinx_run.build()
     # print(sphinx_run.status())
-    assert "Notebook code has no file extension metadata" in sphinx_run.warnings()
     assert "language_info" not in set(sphinx_run.app.env.metadata["basic_unrun"].keys())
     assert sphinx_run.app.env.metadata["basic_unrun"]["author"] == "Chris"